ingresa ruta, devuelve archivos .md y transforma ruta rel a abs
