+++ pruebaRM.md
@@ -1,0 +1,13 @@
+[Pasos para crear una librería en NodeJs](https://medium.com/@anlijudavid/pasos-para-crear-una-librer%C3%ADa-para-nodejs-fa49b17558b8)
+
+[NodeJs para principiantes, aprox 1hr40mins](https://www.youtube.com/watch?v=BhvLIzVL8_o&feature=youtu.be)
+
+[Diferencia visual entre ruta absoluta y relativa](https://www.coffeecup.com/help/articles/absolute-vs-relative-pathslinks/)
+
+
+[Cómo y para qué se usa el fs filesystem](https://www.sohamkamani.com/blog/nodejs-file-system-guide/)
+
+
+[Cómo tomar el directorio actual](https://flaviocopes.com/node-get-current-folder/#:~:text=Using%20.%2F%20will%20give%20you%20the,script%2C%20by%20using%20the%20process.)
+
+[Callbacks](https://github.com/maxogden/art-of-node#callbacks)
